+++ src/main/java/insane96mcp/runeenchanting/data/provider/RERuneTagProvider.java
@@ -56,7 +56,8 @@ public class RERuneTagProvider extends TagsProvider<Rune> {
                 RERunes.CURSE_OF_STATIC_CHARGE,
                 RERunes.CURSE_OF_EXPLOSION,
                 RERunes.CURSE_OF_CORROSION,
-                RERunes.CURSE_OF_UNDEAD);
+                RERunes.CURSE_OF_UNDEAD,
+                RERunes.CURSE_OF_INACCURACY);
 
         add(VANISHABLE,
                 RERunes.CURSE_OF_VANISHING,